+++ tests/core/cbor/FutureTest.php
@@ -9,13 +9,14 @@ class FutureTest extends TestCase
 {
     public function testToString()
     {
-        $future = new Future("time::now()");
-        $this->assertEquals("<future> time::now()", $future->__toString());
+        $future = new Future("{ time::now() }");
+        $this->assertEquals("<future> { time::now() }", $future->__toString());
+        $this->assertEquals("<future> { time::now() }", $future->toString());
     }
 
     public function testJsonSerialize()
     {
-        $future = new Future("time::now()");
-        $this->assertEquals("<future> time::now()", $future->jsonSerialize());
+        $future = new Future("{ time::now() }");
+        $this->assertEquals("<future> { time::now() }", $future->jsonSerialize());
     }
 }
